Worldgen changes - Sandy/gravelly river shores - Fix thin spike blocks breaking on their own randomly - Fix some map colors - Fix "deeply submerged"  trees - Fix the locate command result

+++ src/main/java/net/dries007/tfc/common/blocks/ThinSpikeBlock.java
@@ -150,7 +150,10 @@ public class ThinSpikeBlock extends Block implements IFluidLoggable
     @SuppressWarnings("deprecation")
     public void tick(BlockState state, ServerLevel level, BlockPos pos, RandomSource rand)
     {
-        level.destroyBlock(pos, false);
+        if (!canSurvive(state, level, pos))
+        {
+            level.destroyBlock(pos, false);
+        }
     }
 
     @Override
